Factory device provisioning (#136)
* modify factory device logic and tables

* cleaning up provisioning view

* do not delete factory device

+++ src/main/java/com/nitramite/porssiohjain/services/DeviceService.java
@@ -15,6 +15,7 @@ import com.nitramite.porssiohjain.entity.*;
 import com.nitramite.porssiohjain.entity.enums.AcType;
 import com.nitramite.porssiohjain.entity.enums.DevicePlatform;
 import com.nitramite.porssiohjain.entity.enums.DeviceType;
+import com.nitramite.porssiohjain.entity.enums.FactoryDeviceStatus;
 import com.nitramite.porssiohjain.entity.enums.MqttDeviceProfile;
 import com.nitramite.porssiohjain.entity.enums.ResourceType;
 import com.nitramite.porssiohjain.entity.repository.*;
@@ -155,7 +156,7 @@ public class DeviceService {
     public DeviceResponse getDevice(Long authAccountId, Long deviceId) {
         DeviceEntity device = deviceRepository.findById(deviceId)
                 .orElseThrow(() -> new IllegalArgumentException("Device not found: " + deviceId));
-        if (device.getAccount().getId().equals(authAccountId)) {
+        if (device.getAccount() != null && device.getAccount().getId().equals(authAccountId)) {
             return mapToResponse(device, false);
         } else {
             throw new IllegalStateException("Forbidden!");
@@ -215,7 +216,7 @@ public class DeviceService {
                 .lastCommunication(entity.getLastCommunication())
                 .createdAt(entity.getCreatedAt())
                 .updatedAt(entity.getUpdatedAt())
-                .accountId(entity.getAccount().getId())
+                .accountId(entity.getAccount() != null ? entity.getAccount().getId() : null)
                 .shared(isShared)
                 .apiOnline(entity.isApiOnline())
                 .mqttOnline(entity.isMqttOnline())
@@ -350,6 +351,20 @@ public class DeviceService {
         productionSourceHeatPumpRepository.deleteAll(productionSourceHeatPumpRepository.findByDevice(device));
         weatherControlHeatPumpRepository.deleteAll(weatherControlHeatPumpRepository.findByDevice(device));
         resourceSharingRepository.deleteAll(resourceSharingRepository.findByResourceTypeAndDeviceId(ResourceType.DEVICE, deviceId));
+
+        if (device.getSerialNumber() != null) {
+            device.setAccount(null);
+            device.setDeviceName(device.getSerialNumber());
+            device.setTimezone("UTC");
+            device.setFactoryDeviceStatus(FactoryDeviceStatus.PASSED);
+            device.setClaimedAt(null);
+            device.setEnabled(true);
+            device.setApiOnline(false);
+            device.setMqttOnline(false);
+            deviceRepository.save(device);
+            return;
+        }
+
         deviceRepository.delete(device);
     }
